refactor(torghut): split runtime ledger proof packet assembly
