+++ services/torghut/scripts/assemble_runtime_ledger_proof_packet_modules/__init__.py
@@ -1,0 +1,10 @@
+"""Explicit modules for runtime-ledger proof packet assembly."""
+
+from __future__ import annotations
+
+from scripts.assemble_runtime_ledger_proof_packet_modules.cli import main
+from scripts.assemble_runtime_ledger_proof_packet_modules.packet_builder import (
+    build_runtime_ledger_proof_packet,
+)
+
+__all__ = ("build_runtime_ledger_proof_packet", "main")
